license_server/requirements-render.txt
flask>=2.0
gunicorn
psycopg2-binary
